+++ src/Entity/Config/Contributes/Configuration.ts
@@ -4,6 +4,7 @@ export class Configuration {
      * docomment
      *-----------------------------------------------------------------------*/
     public static KEY_DOCOMMENT = 'docomment';
+    public static SYNTAX = 'syntax';
     public static ACTIVATE_ON_ENTER = 'activateOnEnter';
     public static ADVANCED = 'advanced';
 
@@ -20,10 +21,16 @@ export class Configuration {
     public static INSERT_SPACES = 'insertSpaces';
     public static DETECT_IDENTATION = 'detectIndentation';
 
+    public syntax: CommentSyntax;
     public activateOnEnter: boolean;
     public advanced: Object;
     public eol: string;
     public insertSpaces: boolean;
     public detectIdentation: boolean;
 
+}
+
+export enum CommentSyntax {
+    single = 'single', // three slashes ///
+    delimited = 'delimited', // start with a slash and two stars /**
 }